+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Add: consider `cpsAcceleration` of multiple buying until tier upgrade (X buys to tier amount + buy tier upgrade + Y buys after tier upgrade)
+- Add: consider `cpsAcceleration` of multiple buying until tier upgrade
 
 ## [2042.12] - 2021-09-21
 ### Added